backup/code/server/requirements.txt
flask
opencv-python
face_recognition
mysql-connector-python
